refactor utilities out of Posix
in addition add a readAllFile utility and a new PosixInStr type class
use these throughout the code base where sensible
