+++ src/Plush/Run/BuiltIns/ShellState.hs
@@ -44,6 +44,7 @@ import Plush.Run.BuiltIns.Utilities
 import Plush.Run.BuiltIns.Syntax
 import Plush.Run.Command
 import Plush.Run.Posix
+import Plush.Run.Posix.Utilities
 import Plush.Run.ShellExec
 import Plush.Run.ShellFlags
 import Plush.Run.Types
