+++ src/Plush/Run/ShellExec.hs
@@ -45,6 +45,7 @@ import Data.Maybe (fromMaybe, listToMaybe)
 import qualified Data.Text as T
 
 import Plush.Run.Posix
+import Plush.Run.Posix.Utilities
 import Plush.Run.ShellFlags
 import Plush.Run.Types
 import Plush.Types
